app/requirements.txt
fastapi
uvicorn
langchain>=0.2.0
langchain-community
langchain_huggingface
langchain-ollama
sentence-transformers
faiss-cpu
gtts
deep-translator
PyPDF2
pymongo
python-docx
pandas
python-dotenv
firecrawl
langchain_openai